Separate release notes by major version
