+++ scripts/release/index.ts
@@ -1,14 +1,13 @@
 /* This file is a part of @mdn/browser-compat-data
  * See LICENSE file for more information. */
 
-import fs from 'node:fs/promises';
-
 import chalk from 'chalk-template';
 import esMain from 'es-main';
 
 import { getSemverBumpPulls } from './semver-pulls.js';
-import { getStats, formatStats, Stats } from './stats.js';
-import { getChanges, formatChanges, Changes } from './changes.js';
+import { getStats } from './stats.js';
+import { getChanges } from './changes.js';
+import { getNotes, addNotes } from './notes.js';
 import {
   exec,
   requireGitHubCLI,
@@ -17,61 +16,6 @@ import {
   getRefDate,
   keypress,
 } from './utils.js';
-
-const dirname = new URL('.', import.meta.url);
-
-/**
- * Get the release notes to add
- * @param {string} thisVersion The current version number
- * @param {Changes} changes The changes to format
- * @param {Stats} stats The statistics from the changes
- * @param {string} versionBump Which part of the semver has been bumped
- * @returns {string} The Markdown-formatted release notes
- */
-const getNotes = (
-  thisVersion: string,
-  changes: Changes,
-  stats: Stats,
-  versionBump: string,
-): string =>
-  [
-    `## [${thisVersion}](https://github.com/mdn/browser-compat-data/releases/tag/${thisVersion})`,
-    '',
-    `${new Date().toLocaleDateString('en-US', {
-      month: 'long',
-      day: 'numeric',
-      year: 'numeric',
-    })}`,
-    '',
-    ...(versionBump !== 'patch'
-      ? [
-          '### Notable changes',
-          '',
-          '<!-- TODO: Fill me out with the appropriate information about breaking changes or new backwards-compatible additions! -->',
-          '',
-        ]
-      : []),
-    formatChanges(changes),
-    formatStats(stats),
-  ].join('\n');
-
-/**
- * Add new release notes to the file
- * @param {string} notesToAdd The notes to add to the release notes
- */
-const addNotes = async (notesToAdd: string): Promise<void> => {
-  const notesFilepath = new URL('../../RELEASE_NOTES.md', dirname);
-  const currentNotes = (await fs.readFile(notesFilepath))
-    .toString()
-    .split('\n');
-  const newNotes = [
-    currentNotes[0],
-    currentNotes[1],
-    notesToAdd,
-    ...currentNotes.slice(2),
-  ].join('\n');
-  await fs.writeFile(notesFilepath, newNotes);
-};
 
 /**
  * Perform the commit and submit a pull request
@@ -144,7 +88,7 @@ const main = async () => {
 
   console.log(chalk`{blue Applying changelog...}`);
   const notes = getNotes(thisVersion, changes, stats, versionBump);
-  await addNotes(notes);
+  await addNotes(notes, versionBump, lastVersion);
 
   console.log(chalk`{blue Creating pull request...}`);
   await commitAndPR(thisVersion, versionBump !== 'patch');
